Added function in file adjustidsOut.h. It changes the sign of elements of the idsOut vector in a Block if necessary. The negative sign indicates that the destination Block does not require resources and therefore the walker can be put to the top of the queue, such that it starts before the loop going over the queue is eventually broken due to a lack of available resources. The function should be part of the Group class as soon as we include the vector of Blocks into Group. The function requires that the last Block does not have destination -1 anymore. We dedided that the end is indicated by an id larger than the number of block instead of -1 (because now -1 is a valid id, indicating that we go to Block 1, which does not need resources)

+++ block.cpp
@@ -12,7 +12,7 @@ using namespace std;
 
 Block::Block(int id, vector<int> idsOut, vector<float> probsOut) {
     if ( idsOut.size() == probsOut.size() ) {
-        this->nLegsOut = idsOut.size();
+        this->nLegsOut = (int)idsOut.size();
 
         this->id = id;
         this->idsOut = idsOut;
@@ -56,5 +56,10 @@ mergandBlock::mergandBlock(int id, vector<int> idsOut, vector<float> probsOut, v
         cout << "    DEBUG: block is of type mergandBlock." << endl;
     #endif
 }
+
+
+
+
+
